pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["egomotion"]

[project]
authors = [
    { name = "Giulia D'Angelo" },
    { name = "Alexander Hadjiivanov" },
    { name = "Michael Furlong" },
    { name = "Maria Victoria Clerico" },
]
name = "egomotion"
version = "0.1.0"
requires-python = ">= 3.11"
dependencies = [
    "black",
    "loguru",
    "environs",
    "matplotlib",
    "natsort",
    "nengo-gui",
    "nengo",
    "numpy",
    "opencv-python",
    "pyserial",
    "samna",
    "scikit-image",
    "scipy",
    "sinabs",
    "tonic",
    "torch",
    "torchvision",
    "tqdm",
]
description = "Implementation of an egomotion suppression and attention algorithm on neuromorphic hardware."
readme = "README.md"
license = { file = "LICENCE.md" }

[tool.ruff.format]
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"

[tool.ruff.lint.isort]
from-first = false
force-single-line = true
force-sort-within-sections = true
length-sort = true
lines-after-imports = 1
order-by-type = true
forced-separate = [
    "typing",
    "future",
    "standard-library",
    "third-party",
    "first-party",
    "local-folder",
    "egomotion",
]
relative-imports-order = "closest-to-furthest"
